requirements.txt
streamlit
pip install streamlit-extras
s3fs
pyyaml
st-files-connection
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.3
plotly==5.15.0
pandas==1.5.3
boto3